requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
gunicorn==22.0.0
python-dotenv==1.0.1
